Move page implementations to features and remove module wrappers
