+++ frontend/src/features/system/pages/NotFound.jsx
@@ -1,2 +1,16 @@
-export { default } from '@modules/system/pages/NotFound';
+import React from 'react';
+import { Box, Typography, Button } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
 
+export default function NotFound() {
+  const navigate = useNavigate();
+  return (
+    <Box sx={{ p: 4, textAlign: 'center' }}>
+      <Typography variant="h4" sx={{ mb: 2 }}>Page Not Found</Typography>
+      <Typography color="text.secondary" sx={{ mb: 3 }}>
+        The page you’re looking for doesn’t exist.
+      </Typography>
+      <Button variant="contained" onClick={() => navigate('/')}>Go Home</Button>
+    </Box>
+  );
+}
